MediaControl: Dymically set mainline IVSC media-entity src-pad index
Since mainline kernel commit 48f5fd8967f8 ("media: ivsc: csi: Swap SINK
and SOURCE pads") the src-pad on the mainline ivsc mc-entity is pad 1,
where as on older versions it is pad 0.

When updating the sensor link to point to the IVSC instead, link.srcPad
was left at pad-0 inherited from the original sensor link.

Since with newer kernels this now needs to be pad-1, also update
link.srcPad and do so dynamically, so that it works with all mainline
IVSC driver versions.

This fixes ipu6-camera-hal / icamerasrc not working on laptops with
an IVSC chip running kernels using the mainline IVSC driver with
the mentioned commit.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>